Material local storage design (#18999)
