+++ framework/src/actions/ResolveOptionalMaterialPropertiesAction.C
@@ -31,5 +31,14 @@ void
 ResolveOptionalMaterialPropertiesAction::act()
 {
   mooseAssert(_problem, "Problem doesn't exist");
-  _problem->resolveOptionalMaterialProperties();
+  auto & all_materials = _problem->getMaterialWarehouse();
+
+  for (auto tid : make_range(all_materials.numThreads()))
+    for (auto matbase_ptr : all_materials.getObjects(tid))
+    {
+      auto mat_ptr = std::dynamic_pointer_cast<Material>(matbase_ptr);
+      if (mat_ptr)
+        for (auto & proxy : mat_ptr->getOptionalPropertyProxies())
+          proxy->resolve();
+    }
 }
